Reorganize some tests to fit better with their modules
